web/requirements.txt
flask
flask-socketio
mysql-connector-python
eventlet
werkzeug
redis
